Move Script UI to a separate module and rename it
