+++ crates/sncast/src/starknet_commands/mod.rs
@@ -4,5 +4,5 @@ pub mod declare;
 pub mod deploy;
 pub mod invoke;
 pub mod multicall;
-pub mod script;
 pub mod show_config;
+pub mod script;
