+++ crates/sncast/src/starknet_commands/script/mod.rs
@@ -4,7 +4,7 @@ use std::fs;
 use crate::starknet_commands::declare::BuildConfig;
 use crate::starknet_commands::{call, declare, deploy, invoke};
 use crate::{get_account, get_nonce, WaitForTx};
-use anyhow::{anyhow, ensure, Context, Result};
+use anyhow::{anyhow, Context, ensure, Result};
 use blockifier::execution::common_hints::ExecutionMode;
 use blockifier::execution::deprecated_syscalls::DeprecatedSyscallSelector;
 use blockifier::execution::entry_point::{
@@ -39,80 +39,17 @@ use scarb_api::{package_matches_version_requirement, ScarbCommand};
 use scarb_metadata::Metadata;
 use semver::{Comparator, Op, Version, VersionReq};
 use sncast::helpers::scarb_utils::{
-    get_package_metadata, get_scarb_manifest, get_scarb_metadata_with_deps, CastConfig,
+    CastConfig, get_package_metadata, get_scarb_manifest, get_scarb_metadata_with_deps,
 };
 use sncast::response::print::print_as_warning;
-use sncast::response::print::{print_command_result, OutputFormat, OutputValue};
-use sncast::response::structs::{CommandResponse, ScriptResponse};
-use sncast::NumbersFormat;
+use sncast::response::structs::ScriptResponse;
 use starknet::accounts::Account;
 use starknet::core::types::{BlockId, BlockTag::Pending, FieldElement};
 use starknet::providers::jsonrpc::HttpTransport;
 use starknet::providers::JsonRpcClient;
 use tokio::runtime::Runtime;
 
-#[derive(Copy, Clone, Debug, Default, Eq, PartialEq, Ord, PartialOrd)]
-pub enum Verbosity {
-    /// Silence all script output except for errors
-    Quiet,
-
-    /// Default verbosity level
-    #[default]
-    Normal,
-}
-
-#[derive(Debug)]
-pub struct UI {
-    verbosity: Verbosity,
-    numbers_format: NumbersFormat,
-    output_format: OutputFormat,
-}
-
-impl UI {
-    pub fn new(is_quiet: bool, numbers_format: NumbersFormat, output_format: OutputFormat) -> Self {
-        let verbosity = if is_quiet {
-            Verbosity::Quiet
-        } else {
-            Verbosity::Normal
-        };
-
-        Self {
-            verbosity,
-            numbers_format,
-            output_format,
-        }
-    }
-
-    pub fn is_quiet(&self) -> bool {
-        self.verbosity < Verbosity::Normal
-    }
-
-    pub fn is_json(&self) -> bool {
-        self.output_format == OutputFormat::Json
-    }
-
-    pub fn print_subcommand_response<T: CommandResponse>(
-        &self,
-        command: &str,
-        response: T,
-    ) -> Result<()> {
-        if self.verbosity >= Verbosity::Normal {
-            let header = (
-                String::from("script_subcommand"),
-                OutputValue::String(command.to_string()),
-            );
-            print_command_result(
-                header,
-                &mut Ok(response),
-                self.numbers_format,
-                &self.output_format,
-            )?;
-            println!();
-            return Ok(());
-        }
-        Ok(())
-    }
-}
+pub mod logger;
 
 #[derive(Args)]
 #[command(about = "Execute a deployment script")]
@@ -130,7 +67,7 @@ pub struct CastScriptExtension<'a> {
     pub provider: &'a JsonRpcClient<HttpTransport>,
     pub tokio_runtime: Runtime,
     pub config: &'a CastConfig,
-    pub script_ui: UI,
+    pub script_ui: logger::ScriptLogger,
 }
 
 impl<'a> ExtensionLogic for CastScriptExtension<'a> {
@@ -343,7 +280,7 @@ pub fn run(
     provider: &JsonRpcClient<HttpTransport>,
     tokio_runtime: Runtime,
     config: &CastConfig,
-    script_ui: UI,
+    script_ui: logger::ScriptLogger,
 ) -> Result<ScriptResponse> {
     let path = compile_script(path_to_scarb_toml.clone(), &script_ui)?;
 
@@ -473,7 +410,7 @@ fn warn_if_sncast_std_not_compatible(scarb_metadata: &Metadata) -> Result<()> {
     Ok(())
 }
 
-fn compile_script(path_to_scarb_toml: Option<Utf8PathBuf>, script_ui: &UI) -> Result<Utf8PathBuf> {
+fn compile_script(path_to_scarb_toml: Option<Utf8PathBuf>, script_ui: &logger::ScriptLogger) -> Result<Utf8PathBuf> {
     let scripts_manifest_path = path_to_scarb_toml.unwrap_or_else(|| {
         get_scarb_manifest()
             .context("Failed to retrieve manifest path from scarb")
